requirements.txt
seaborn
tensorflow
scikit-learn
matplotlib
numpy
pillow
pandas
tkinter
reportlab
